+++ test/integration/Setup-Scenario9.ps1
@@ -144,7 +144,6 @@ try {
     $usernameSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Username" }
     $passwordSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Password" }
     $useSSLSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Use Secure Connection (LDAPS)?" }
-    $certValidationSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Certificate Validation" }
     $connectionTimeoutSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Connection Timeout" }
     $authTypeSetting = $ldapConnectorFull.settings | Where-Object { $_.name -eq "Authentication Type" }
 
@@ -154,9 +153,6 @@ try {
     if ($usernameSetting) { $ldapSettings[$usernameSetting.id] = @{ stringValue = $DirectoryConfig.BindDN } }
     if ($passwordSetting) { $ldapSettings[$passwordSetting.id] = @{ stringValue = $DirectoryConfig.BindPassword } }
     if ($useSSLSetting) { $ldapSettings[$useSSLSetting.id] = @{ checkboxValue = $DirectoryConfig.UseSSL } }
-    if ($certValidationSetting -and $DirectoryConfig.CertValidation) {
-        $ldapSettings[$certValidationSetting.id] = @{ stringValue = $DirectoryConfig.CertValidation }
-    }
     if ($connectionTimeoutSetting) { $ldapSettings[$connectionTimeoutSetting.id] = @{ intValue = 30 } }
     if ($authTypeSetting) { $ldapSettings[$authTypeSetting.id] = @{ stringValue = $DirectoryConfig.AuthType } }